+++ src/main/java/hotspot/worker/common/config/kafka/producer/KafkaProducerConfig.java
@@ -7,7 +7,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.kafka.core.ProducerFactory;
 
-import hotspot.worker.consumer.schema.UsageAlertEvent;
+import hotspot.worker.consumer.usage.schema.UsageAlertEvent;
 import hotspot.worker.producer.schema.UsageEvent;
 
 @Configuration
